pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "customer_support"
version = "0.1.0"
description = "RAG-based customer support AI with DDD architecture"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-ulid>=2.2.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0",
    "faiss-cpu>=1.8.0",
]
azure = [
    "azure-storage-blob>=12.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["src"]

# ── RUFF ──────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.per-file-ignores]
"**/prompts.py" = ["E501"]  # Prompt strings must not be wrapped

[tool.ruff.format]
quote-style = "double"

# ── PYTEST ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: pure domain logic, no I/O, runs in milliseconds",
    "integration: pipeline with mocks, no real external services",
    "e2e: real external services, run explicitly",
]

# ── COVERAGE ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["src"]
omit = [
    "src/customer_support/main.py",
    "src/customer_support/retrieval/store_builder.py",
    "src/customer_support/*/prompts.py",
]

[tool.coverage.report]
show_missing = true
# 80% reflects that store_builder.py and prompts are excluded,
# and that pure functions in checker.py are straightforward to test.
fail_under = 80

# ── MYPY ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true